+++ sample-apps/react/react-dogfood/components/ActiveCall.tsx
@@ -59,6 +59,11 @@ import { StepNames, useTourContext } from '../context/TourContext';
 import { useNotificationSounds } from '../hooks/useNotificationSounds';
 import { usePipWindow } from '../hooks/usePipWindow';
 import { StagePip } from './StagePip';
+import {
+  RemoteVideoControls,
+  useRemoteFilePublisher,
+} from './RemoteFilePublisher';
+import { ModerationNotification } from './ModerationNotification';
 
 export type ActiveCallProps = {
   chatClient?: StreamChat | null;
@@ -152,6 +157,8 @@ export const ActiveCall = (props: ActiveCallProps) => {
   } = usePipWindow('@pronto/pip');
   useNotificationSounds();
 
+  const remoteFilePublisherAPI = useRemoteFilePublisher();
+
   return (
     <div className="rd__call">
       {isDemoEnvironment && <TourPanel highlightClass="rd__highlight" />}
@@ -230,6 +237,7 @@ export const ActiveCall = (props: ActiveCallProps) => {
             )}
           </div>
         </div>
+        <ModerationNotification />
         <div className="rd__notifications">
           <Restricted
             requiredGrants={[OwnCapability.SEND_AUDIO]}
@@ -273,6 +281,10 @@ export const ActiveCall = (props: ActiveCallProps) => {
             </div>
           </div>
           <div className="str-video__call-controls--group str-video__call-controls--media">
+            {remoteFilePublisherAPI && (
+              <RemoteVideoControls api={remoteFilePublisherAPI} />
+            )}
+
             <ToggleDualMicButton />
             <ToggleDualCameraButton />
             <div className="str-video__call-controls__desktop">